rename module bench to avodi conflict
